pyproject.toml
[tool.poetry]
name = "evok2mqtt"
version = "0.1.0"
description = "Bridge between EVOK unipi interface and MQTT"
authors = ["Martijn Hemeryck <martijn.hemeryck@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
websockets = "^8.1"
paho-mqtt = "^1.5"
pyyaml = "^5.1"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
isort = "^4.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
